# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "app队列管理服务发服务",
        "title": "QueueMgrServ",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:81",
    "basePath": "/queue_mgr",
    "paths": {
        "/bind": {
            "post": {
                "description": "bind queue consumer",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QueueMgrServ"
                ],
                "summary": "给队列绑定消费协程",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "consumer/消费器名",
                        "name": "consumer",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "queue/队列名",
                        "name": "queue",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "properties/绑定消费器相关参数(json)",
                        "name": "properties",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    }
                }
            }
        },
        "/consumer/create": {
            "post": {
                "description": "create queue consumer",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QueueMgrServ"
                ],
                "summary": "创建队列消费器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "consumer/消费器名",
                        "name": "consumer",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "\"FastCGI\"",
                            "\"Native\"",
                            "\"Shell\"",
                            "\"Api\"",
                            "\"Grpc\"",
                            "\"Proxy\"",
                            "\"Plugins\""
                        ],
                        "type": "string",
                        "description": "type/消费器类型",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "properties/消费器相关参数(json)",
                        "name": "properties",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    }
                }
            }
        },
        "/consumers": {
            "get": {
                "description": "query queue consumer lists",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QueueMgrServ"
                ],
                "summary": "罗列消费器列信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page/页码",
                        "name": "page",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "count/分页量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "state/消费进程状态",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"created_at:desc\"",
                        "description": "sort/排序参数",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name/限定队列名(模糊匹配eg: test*)",
                        "name": "queue",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name/限定消费器名(模糊匹配eg: test*)",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "description": "prometheus Metrics",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "QueueMgrServ"
                ],
                "summary": "export prometheus metrics"
            }
        },
        "/queue/create": {
            "post": {
                "description": "add queue  info",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QueueMgrServ"
                ],
                "summary": "创建可消费队列信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "queue/队列名",
                        "name": "queue",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "\"AMQP\"",
                            "\"MQTT\"",
                            "\"HTTP\"",
                            "\"WS\"",
                            "\"PLUGINS\""
                        ],
                        "type": "string",
                        "description": "queue/队列链接驱动器类型",
                        "name": "driver",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "properties/可消费队列相关参数(json)",
                        "name": "properties",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    }
                }
            }
        },
        "/queues": {
            "get": {
                "description": "query queue consumer lists",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QueueMgrServ"
                ],
                "summary": "罗列消费队列信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page/页码",
                        "name": "page",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "count/分页量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "state/消费进程状态",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"created_at:desc\"",
                        "description": "sort/排序参数",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name/限定队列名(模糊匹配eg: test*)",
                        "name": "queue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    }
                }
            }
        },
        "/routers": {
            "get": {
                "description": "List Service Api Routers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QueueMgrServ"
                ],
                "summary": "罗列服务接口列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.KvMap"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.KvMap"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.KvMap"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.KvMap"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/entity.KvMap"
                        }
                    }
                }
            }
        },
        "/state": {
            "get": {
                "description": "query queue consumer state",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QueueMgrServ"
                ],
                "summary": "查询队列消费器状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "queue/消费队列名",
                        "name": "queue",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "tag/消费进程标签",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "state/消费进程状态",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    }
                }
            }
        },
        "/state/update": {
            "post": {
                "description": "change queue consumer state or processor number",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QueueMgrServ"
                ],
                "summary": "控制队列消费器状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "queue/消费队列名",
                        "name": "queue",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "state/消费进程状态",
                        "name": "state",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "tag/消费进程标签",
                        "name": "tag",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "scale/消费队列协程数扩缩容",
                        "name": "scale",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.JsonData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "info": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.KvMap"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "entity.JsonResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data",
                    "$ref": "#/definitions/entity.JsonData"
                },
                "ret": {
                    "description": "httpCode",
                    "type": "integer"
                }
            }
        },
        "entity.KvMap": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "read": " Grants read access",
                "write": " Grants write access"
            }
        }
    },
    "tags": [
        {
            "description": "app队列管理服务发服务接口集合 服务路由前缀: /queue_mgr/, jwt scope: queueMgrServ",
            "name": "QueueMgrServ"
        }
    ],
    "x-extension-openapi": {
        "name": "QueueMgrServ"
    }
}